pyproject.toml
[project]
name = "explain-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.3.25",
    "langchain-core>=0.3.58",
    "langchain-google-genai>=2.1.4",
    "langchain-mcp-adapters>=0.0.10",
    "langchain-ollama>=0.3.2",
    "langchain-openai>=0.3.16",
    "langgraph>=0.4.3",
    "mcp[cli]>=1.7.1",
    "python-dotenv>=1.1.0",
]
